java/live-study/4. 제어문/source/src/test/java/org/java/study/assignment5/QueueTest.java
package org.java.study.assignment5;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.*;

class QueueTest {

    @Test
    void test() {
        Queue queue = new Queue(3);

        queue.enqueue(1);
        queue.enqueue(2);
        queue.enqueue(3);

        assertEquals(3, queue.size());
        assertTrue(queue.isFull());
        assertThrows(IndexOutOfBoundsException.class, () -> queue.enqueue(4));

        assertEquals(1, queue.peek());
        assertEquals(1, queue.dequeue());
        assertEquals(2, queue.dequeue());
        assertEquals(3, queue.dequeue());
        assertTrue(queue.isEmpty());

        queue.enqueue(1);
        assertEquals(1, queue.size());
    }

}